[IMP] estate: website task completed
PURPOSE

to give user functionality of Filtering the properties <u>Listed After</u> date.

to give user a breadcrumb for ease of access while surfing around the website.

Specifications

Added a Form with a single input and onclick of submit redirected it to the
controller with the date added in post and appended a DOMAIN of CREATE_DATE to
the domain.

Added a simple JS breadcrumb with static link as only 2 pages are there to
surf around the website.

LINKS

PR  #00001
Task 000001